Both skill trees zoomable and pannable
