+++ package.json
@@ -19,6 +19,7 @@
     "@types/react": "18.0.26",
     "@types/react-dom": "18.0.10",
     "@vercel/og": "0.0.27",
+    "d3": "^7.8.2",
     "eslint": "8.30.0",
     "eslint-config-next": "13.1.1",
     "flowbite": "1.6.3",
@@ -50,6 +51,7 @@
     "@types/lodash": "4.14.191",
     "autoprefixer": "10.4.13",
     "postcss": "8.4.21",
-    "tailwindcss": "3.2.4"
+    "tailwindcss": "3.2.4",
+    "@types/d3": "^7.4.0"
   }
 }
